+++ views/admin/_profile.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @var yii\widgets\ActiveForm       $form
+ * @var dektrium\user\models\Profile $profile
+ */
+
+?>
+
+<?= $form->field($profile, 'name') ?>
+<?= $form->field($profile, 'public_email') ?>
+<?= $form->field($profile, 'website') ?>
+<?= $form->field($profile, 'location') ?>
+<?= $form->field($profile, 'gravatar_email') ?>
+<?= $form->field($profile, 'bio')->textarea() ?>