+++ views/admin/create.php
@@ -13,8 +13,8 @@ use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 
 /**
- * @var yii\web\View $this
- * @var dektrium\user\models\User $model
+ * @var yii\web\View              $this
+ * @var dektrium\user\models\User $user
  */
 
 $this->title = Yii::t('user', 'Create a user account');
@@ -30,16 +30,15 @@ $this->params['breadcrumbs'][] = $this->title;
     </div>
     <div class="panel-body">
         <div class="alert alert-info">
-            <?= Yii::t('user', 'Password and username will be sent to user by email') ?>.
-            <?= Yii::t('user', 'If you want password to be generated automatically leave its field empty') ?>.
+            <?= Yii::t('user', 'Credentials will be sent to user by email') ?>.
+            <?= Yii::t('user', 'If you want to be generate password automatically leave password field empty') ?>.
         </div>
-        <?php $form = ActiveForm::begin(); ?>
+        <?php $form = ActiveForm::begin([
+            'enableAjaxValidation'   => true,
+            'enableClientValidation' => false
+        ]); ?>
 
-        <?= $form->field($model, 'username')->textInput(['maxlength' => 25, 'autofocus' => true]) ?>
-
-        <?= $form->field($model, 'email')->textInput(['maxlength' => 255]) ?>
-
-        <?= $form->field($model, 'password')->passwordInput() ?>
+        <?= $this->render('_user', ['form' => $form, 'user' => $user]) ?>
 
         <div class="form-group">
             <?= Html::submitButton(Yii::t('user', 'Save'), ['class' => 'btn btn-success']) ?>