+++ views/admin/update.php
@@ -9,44 +9,52 @@
  * file that was distributed with this source code.
  */
 
-
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 
 /**
- * @var yii\web\View $this
- * @var dektrium\user\models\User $model
+ * @var yii\web\View                 $this
+ * @var dektrium\user\models\User    $user
+ * @var dektrium\user\models\Profile $profile
+ * @var dektrium\user\Module         $module
  */
 
 $this->title = Yii::t('user', 'Update user account');
 $this->params['breadcrumbs'][] = ['label' => Yii::t('user', 'Users'), 'url' => ['index']];
 $this->params['breadcrumbs'][] = $this->title;
 ?>
-<h1><i class="glyphicon glyphicon-user"></i> <?= Html::encode($model->username) ?>
-    <?php if (!$model->getIsConfirmed()): ?>
-        <?= Html::a(Yii::t('user', 'Confirm'), ['confirm', 'id' => $model->id], ['class' => 'btn btn-success btn-xs', 'data-method' => 'post']) ?>
-    <?php endif; ?>
-    <?php if ($model->getIsBlocked()): ?>
-        <?= Html::a(Yii::t('user', 'Unblock'), ['block', 'id' => $model->id], ['class' => 'btn btn-success btn-xs', 'data-method' => 'post', 'data-confirm' => Yii::t('user', 'Are you sure to block this user?')]) ?>
-    <?php else: ?>
-        <?= Html::a(Yii::t('user', 'Block'), ['block', 'id' => $model->id], ['class' => 'btn btn-danger btn-xs', 'data-method' => 'post', 'data-confirm' => Yii::t('user', 'Are you sure to block this user?')]) ?>
-    <?php endif; ?>
-</h1>
 
 <?php echo $this->render('flash') ?>
+<div class="alert alert-info">
+    <?= Yii::t('user', 'Registered at {0, date, MMMM dd, YYYY HH:mm} from {1}', [$user->created_at, is_null($user->registration_ip) ? 'N/D' : long2ip($user->registration_ip)]) ?>
+</div>
+<?php if ($module->enableConfirmation && $user->getIsConfirmed()): ?>
+    <div class="alert alert-success">
+        <?= Yii::t('user', 'Confirmed at {0, date, MMMM dd, YYYY HH:mm}', [$user->created_at]) ?>
+    </div>
+<?php endif; ?>
+<?php if ($user->getIsBlocked()): ?>
+    <div class="alert alert-danger">
+        <?= Yii::t('user', 'Blocked at {0, date, MMMM dd, YYYY HH:mm}', [$user->blocked_at]) ?>
+    </div>
+<?php endif;?>
 
-<div class="panel panel-info">
-    <div class="panel-heading"><?= Yii::t('user', 'Information') ?></div>
+<?php $form = ActiveForm::begin([
+    'enableAjaxValidation'   => true,
+    'enableClientValidation' => false
+]); ?>
+
+<div class="panel panel-default">
     <div class="panel-body">
-        <?= Yii::t('user', 'Registered at {0, date, MMMM dd, YYYY HH:mm} from {1}', [$model->created_at, is_null($model->registration_ip) ? 'N/D' : long2ip($model->registration_ip)]) ?>
-        <br/>
-        <?php if (Yii::$app->getModule('user')->enableConfirmation && $model->getIsConfirmed()): ?>
-            <?= Yii::t('user', 'Confirmed at {0, date, MMMM dd, YYYY HH:mm}', [$model->created_at]) ?>
-            <br/>
+        <?= Html::submitButton(Yii::t('user', 'Save'), ['class' => 'btn btn-primary btn-sm']) ?>
+        <?php if (!$user->getIsConfirmed()): ?>
+            <?= Html::a(Yii::t('user', 'Confirm'), ['confirm', 'id' => $user->id, 'back' => 'update'], ['class' => 'btn btn-success btn-sm', 'data-method' => 'post']) ?>
         <?php endif; ?>
-        <?php if ($model->getIsBlocked()): ?>
-            <?= Yii::t('user', 'Blocked at {0, date, MMMM dd, YYYY HH:mm}', [$model->blocked_at]) ?>
-        <?php endif;?>
+        <?php if ($user->getIsBlocked()): ?>
+            <?= Html::a(Yii::t('user', 'Unblock'), ['block', 'id' => $user->id, 'back' => 'update'], ['class' => 'btn btn-success btn-sm', 'data-method' => 'post', 'data-confirm' => Yii::t('user', 'Are you sure to block this user?')]) ?>
+        <?php else: ?>
+            <?= Html::a(Yii::t('user', 'Block'), ['block', 'id' => $user->id, 'back' => 'update'], ['class' => 'btn btn-danger btn-sm', 'data-method' => 'post', 'data-confirm' => Yii::t('user', 'Are you sure to block this user?')]) ?>
+        <?php endif; ?>
     </div>
 </div>
 
@@ -55,18 +63,31 @@ $this->params['breadcrumbs'][] = $this->title;
         <?= Html::encode($this->title) ?>
     </div>
     <div class="panel-body">
-        <?php $form = ActiveForm::begin(); ?>
-
-        <?= $form->field($model, 'username')->textInput(['maxlength' => 25]) ?>
-
-        <?= $form->field($model, 'email')->textInput(['maxlength' => 255]) ?>
-
-        <?= $form->field($model, 'password')->passwordInput() ?>
-
-        <div class="form-group">
-            <?= Html::submitButton(Yii::t('user', 'Save'), ['class' => 'btn btn-primary']) ?>
-        </div>
-
-        <?php ActiveForm::end(); ?>
+        <?= $this->render('_user', ['form' => $form, 'user' => $user]) ?>
     </div>
 </div>
+
+<div class="panel panel-default">
+    <div class="panel-heading">
+        <?= Yii::t('user', 'Update user profile') ?>
+    </div>
+    <div class="panel-body">
+        <?= $this->render('_profile', ['form' => $form, 'profile' => $profile]) ?>
+    </div>
+</div>
+
+<div class="panel panel-default">
+    <div class="panel-body">
+        <?= Html::submitButton(Yii::t('user', 'Save'), ['class' => 'btn btn-primary btn-sm']) ?>
+        <?php if (!$user->getIsConfirmed()): ?>
+            <?= Html::a(Yii::t('user', 'Confirm'), ['confirm', 'id' => $user->id, 'back' => 'update'], ['class' => 'btn btn-success btn-sm', 'data-method' => 'post']) ?>
+        <?php endif; ?>
+        <?php if ($user->getIsBlocked()): ?>
+            <?= Html::a(Yii::t('user', 'Unblock'), ['block', 'id' => $user->id, 'back' => 'update'], ['class' => 'btn btn-success btn-sm', 'data-method' => 'post', 'data-confirm' => Yii::t('user', 'Are you sure to block this user?')]) ?>
+        <?php else: ?>
+            <?= Html::a(Yii::t('user', 'Block'), ['block', 'id' => $user->id, 'back' => 'update'], ['class' => 'btn btn-danger btn-sm', 'data-method' => 'post', 'data-confirm' => Yii::t('user', 'Are you sure to block this user?')]) ?>
+        <?php endif; ?>
+    </div>
+</div>
+
+<?php ActiveForm::end(); ?>